# To_Do_host/bin/Debug/net6.0/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "To_Do_wpfweb, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/To_Do_": {
      "get": {
        "tags": [
          "To_Do_"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/To_Do"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/To_Do"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/To_Do"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "To_Do_"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/To_Do"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/To_Do"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/To_Do"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/To_Do_/{id}": {
      "get": {
        "tags": [
          "To_Do_"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/To_Do"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/To_Do"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/To_Do"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "To_Do_"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "To_Do": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "plan": {
            "type": "string",
            "nullable": true
          },
          "titleId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}